refactor: 공통 컴포넌트 @common 디렉토리로 이동
