+++ frontend/src/components/@common/Box/Box.stories.js
