+++ frontend/src/components/@common/Label/Label.stories.js
