--- frontend/src/components/form/index.js
@@ -1,11 +1,0 @@
-export { default as Form } from "./Form/Form";
-export { default as Label } from "./Label/Label";
-export { default as Radio } from "./Radio/Radio";
-export { default as CheckBox } from "./CheckBox/CheckBox";
-export { default as Button } from "./Button/Button";
-export { default as Description } from "./Description/Description";
-export { default as TextInput } from "./TextInput/TextInput";
-export { default as TextInputField } from "./TextInputField/TextInputField";
-export { default as SummaryCheckField } from "./SummaryCheckField/SummaryCheckField";
-export { default as GenderField } from "./GenderField/GenderField";
-export { default as BirthField } from "./BirthField/BirthField";
