+++ frontend/src/provider/FormProvider/FormInput.js
@@ -1,7 +1,7 @@
 import { useEffect } from "react";
-import { TextInputField } from "../../components/form";
 import useFormContext from "../../hooks/useFormContext";
 import PropTypes from "prop-types";
+import TextInputField from "../../components/form/TextInputField/TextInputField";
 
 const FormInput = ({ name, initialValue, ...props }) => {
   const { value, errorMessage, handleChange, register, unRegister } =
