+++ frontend/src/provider/FormProvider/SubmitButton.js
@@ -1,6 +1,6 @@
-import { Button } from "../../components/form";
 import useFormContext from "../../hooks/useFormContext";
 import PropTypes from "prop-types";
+import Button from "../../components/@common/Button/Button";
 
 const SubmitButton = ({ children }) => {
   const { isValid, isEmpty } = useFormContext();
